NetworkSearchAgent/requirements.txt
langchain-openai>=1.0.2
langchain-core>=0.3.0
tavily-python>=0.5.0
deepagents>=0.2.6
python-dotenv>=1.0.0
langgraph-cli[inmem]>=0.1.55